requirements.txt
scikit-learn==1.3.2
numpy==1.24.4
pandas>=1.3.0,<1.6.0
joblib==1.3.1
pycaret==3.0.0
kaggle==1.5.12
apache-airflow==2.9.2